Added code for issue #85 and a lot of refactoring surrounding it
